carrie silly air duster

+++ include/drivetrain/movement.hpp
@@ -1,5 +1,5 @@
 #include "main.h"
 
-extern void setVels(int linVel, int angVel);
+extern void setVels(float linVel, float angVel);
 extern void simpleDrive(int y, int turn);
 extern void MPDrive(const std::string filePath);
